Fix issue with program freezing during permute method
The last step of the permute method (output_final_sublists) would iterate through all of the sorted sublist options, and shovel lists containing no duplicates into the final array (sublists). The bug presented itself when the the remaining sublist options contained duplicates before the number of sublists reached the number of groups. This caused the until loop condition to run infinitely.

This commit removes the until loop, and checks afterward if we have more sublists than desired groups. If the result is invalid, it is set to nil, and removed before the comparison step. There is not currently a test for when the sublist needs to be nullified, it is unlikely that this result would be the final result returned.